syscalls

+++ userland/mlibc-patch/abis/modulos/pid_t.h
@@ -1,7 +1,8 @@
+
 #ifndef _ABIBITS_PID_T_H
 #define _ABIBITS_PID_T_H
 
-typedef unsigned long pid_t;
+typedef int pid_t;
 
 #endif /* _ABIBITS_PID_T_H */
